+++ src/params.rs
@@ -58,13 +58,10 @@
 //!
 #![doc = include_str!("docs/command.md")]
 //!
-use std::{ffi::OsString, marker::PhantomData};
+use std::{ffi::OsString, marker::PhantomData, str::FromStr};
 
-use super::{Args, Error, OptionParser, Parser};
-use crate::{args::Arg, from_os_str::FromOsStr, item::ShortLong, Item, Meta};
-
-#[cfg(doc)]
-use crate::from_os_str::FromUtf8;
+use super::{parse_os_str, Args, Error, OptionParser, Parser};
+use crate::{args::Arg, item::ShortLong, Item, Meta};
 
 /// A named thing used to create [`flag`](NamedArg::flag), [`switch`](NamedArg::switch) or
 /// [`argument`](NamedArg::argument)
@@ -353,8 +350,9 @@ impl NamedArg {
     /// literal can't start with `-` unless separated from the flag with `=`. For short flags value
     /// can follow immediately: `-fbar`.
     ///
-    /// When using combinatoring API you must specify the type with turbofish, for parsing types
-    /// that don't implement [`FromOsStr`] you can use [`FromUtf8`] helper tag.
+    /// When using combinatoring API you can specify the type with turbofish, for parsing types
+    /// that don't implement [`FromStr`] you can use consume a `String`/`OsString` first and parse
+    /// it by hands.
     /// ```rust
     /// # use bpaf::*;
     /// fn parse_arg() -> impl Parser<usize> {
@@ -365,7 +363,7 @@ impl NamedArg {
     #[must_use]
     pub fn argument<T>(self, metavar: &'static str) -> ParseArgument<T>
     where
-        T: FromOsStr,
+        T: FromStr + 'static,
     {
         build_argument(self, metavar)
     }
@@ -389,8 +387,9 @@ impl NamedArg {
 /// `cat hello world` and `cat world hello` would display contents of the same two files but in
 /// different order.
 ///
-/// When using combinatoring API you must specify the type with turbofish, for parsing types
-/// that don't implement [`FromOsStr`] you can use [`FromUtf8`] helper tag.
+/// When using combinatoring API you can specify the type with turbofish, for parsing types
+/// that don't implement [`FromStr`] you can use consume a `String`/`OsString` first and parse
+/// it by hands.
 /// ```no_run
 /// # use bpaf::*;
 /// fn parse_pos() -> impl Parser<usize> {
@@ -722,13 +721,14 @@ impl<T> ParseArgument<T> {
     }
 }
 
-impl<O, T> Parser<O> for ParseArgument<T>
+impl<T> Parser<T> for ParseArgument<T>
 where
-    T: FromOsStr<Out = O>,
+    T: FromStr + 'static,
+    <T as std::str::FromStr>::Err: std::fmt::Display,
 {
-    fn eval(&self, args: &mut Args) -> Result<O, Error> {
+    fn eval(&self, args: &mut Args) -> Result<T, Error> {
         let os = self.take_argument(args)?;
-        match T::from_os_str(os) {
+        match parse_os_str::<T>(os) {
             Ok(ok) => Ok(ok),
             Err(err) => Err(args.word_parse_error(&err)),
         }
@@ -901,13 +901,14 @@ fn parse_word(
     }
 }
 
-impl<O, T> Parser<O> for ParsePositional<T>
+impl<T> Parser<T> for ParsePositional<T>
 where
-    T: FromOsStr<Out = O>,
+    T: FromStr + 'static,
+    <T as std::str::FromStr>::Err: std::fmt::Display,
 {
-    fn eval(&self, args: &mut Args) -> Result<O, Error> {
+    fn eval(&self, args: &mut Args) -> Result<T, Error> {
         let os = parse_word(args, self.strict, self.metavar, &self.help)?;
-        match T::from_os_str(os) {
+        match parse_os_str::<T>(os) {
             Ok(ok) => Ok(ok),
             Err(err) => Err(args.word_parse_error(&err)),
         }
@@ -930,10 +931,12 @@ pub struct ParseAny<T> {
 ///
 /// `any` behaves similar to [`positional`] so you should be using it near the right most end of
 /// the consumer struct. Note, consuming "anything" also consumes `--help` unless restricted
-/// with `guard`
+/// with `guard`. It's better stick to `positional` unless you are trying to consume raw options
+/// to pass to some other process or do some special handling.
 ///
-/// When using combinatoring API you must specify the type with turbofish, for parsing types
-/// that don't implement [`FromOsStr`] you can use [`FromUtf8`] helper tag.
+/// When using combinatoring API you can specify the type with turbofish, for parsing types
+/// that don't implement [`FromStr`] you can use consume a `String`/`OsString` first and parse
+/// it by hands. For `any` you would usually consume it either as a `String` or `OsString`.
 /// ```no_run
 /// # use bpaf::*;
 /// # use std::ffi::OsString;
@@ -1007,13 +1010,14 @@ impl<T> ParseAny<T> {
     }
 }
 
-impl<T, O> Parser<O> for ParseAny<T>
+impl<T> Parser<T> for ParseAny<T>
 where
-    T: FromOsStr<Out = O>,
+    T: FromStr + 'static,
+    <T as std::str::FromStr>::Err: std::fmt::Display,
 {
-    fn eval(&self, args: &mut Args) -> Result<O, Error> {
+    fn eval(&self, args: &mut Args) -> Result<T, Error> {
         let os = self.next_os_string(args)?;
-        match T::from_os_str(os) {
+        match parse_os_str::<T>(os) {
             Ok(ok) => Ok(ok),
             Err(err) => Err(args.word_parse_error(&err)), // Error::Stderr(err)),
         }